+++ src/pages/ErrorPage/ErrorPage.test.tsx
@@ -20,6 +20,6 @@ describe("ErrorPage", () => {
     const { getByText } = customRender(<ErrorPage />);
     const home = getByText("Home");
     await userEvent.click(home);
-    expect(window.location.pathname).toEqual("/");
+    expect(window.location.pathname).toEqual("/auth");
   });
 });
