+++ src/pages/Login/Login.test.tsx
@@ -3,18 +3,16 @@ import { customRender } from "../../utils/testUtils";
 import Login from "./Login";
 
 describe("Should pass all login tests", () => {
-  // This is a NOP function
-  const fakeSetState = () => null 
 
   it("should render welcome title on the page", () => {
-    customRender(<Login setUserUID={fakeSetState} />);
+    customRender(<Login />);
 
     const welcomeTitle = screen.getByText(/Welcome Back/i);
     expect(welcomeTitle).toBeInTheDocument();
   });
 
   it("should render form for email and password", () => {
-    customRender(<Login setUserUID={fakeSetState} />);
+    customRender(<Login />);
     const emailInputText = screen.getByPlaceholderText(/you@example\.com/i);
     expect(emailInputText).toBeInTheDocument();
   });
